% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmpPower.R
\name{calcEmpPower}
\alias{calcEmpPower}
\title{Calculate Empirical Power}
\usage{
calcEmpPower(n, delta, nboot, alpha = 0.05, bonferroni = 1129)
}
\arguments{
\item{n}{Integer. Sample size.}

\item{delta}{Numeric. Effect size.}

\item{nboot}{Integer. Number of simulations to perform
in estimating \code{1 - beta}.}

\item{alpha}{Numeric. The significance threshold.}

\item{bonferroni}{Numeric. Size of the Bonferroni correction to adjust for.}
}
\value{
The value of the empirically calculated power (\code{1 - beta}).
}
\description{
Simulate the empirical power of a test through simulated data
with a given effect size of sample size (per group).
}
\examples{
calcEmpPower(10, 15, 1.5)
}
\seealso{
\code{\link[=t.test]{t.test()}}, \code{\link[=qt]{qt()}}, \code{\link[=rnorm]{rnorm()}}
}
\author{
Stu Field
}
