% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{cortar}
\alias{cortar}
\title{Convert numeric to character}
\usage{
cortar(x)
}
\arguments{
\item{x}{numeric vector which is to be converted to a character vector by
cutting.}

\item{breaks}{either a numeric vector of two or more unique cut points or a
single number (greater than or equal to 2) giving the number of intervals
into which x is to be cut.}

\item{open_left}{logical, \code{TRUE} indicates to include in the interval the
left limit}

\item{open_right}{logica, \code{TRUE} indicates to include in the interval the
right limit}

\item{dig_lab}{nteger which is used when labels are not given. It determines
the number of digits used in formatting the break numbers.}
}
\value{
a character vector with the new codification
}
\description{
\code{cortar} divides the range of x into intervals and codes
the values in x according to which interval they fall. The leftmost
interval corresponds to level one, the next leftmost to level two and so on.
}
\details{
\code{cortar} is similar to \code{base::cut}, but we have modify to use
with the package \code{bookdown}
}
\examples{
x <- rnorm(1000)
cortar(x, breaks = -10:10)
}
