% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-power-table.R
\name{create_power_table}
\alias{create_power_table}
\title{Calculate Power Table}
\usage{
create_power_table(
  power = seq(0.6, 0.95, 0.05),
  n_vec = seq(20, 100, 10),
  bonferroni = 1000
)
}
\arguments{
\item{power}{\code{numeric(n)}. Which level(s) of power to
evaluate the required samples size.}

\item{n_vec}{\code{integer(n)}. Which sample sizes per group
to evaluate the corresponding power.}

\item{bonferroni}{\code{integer(1)}. The Bonferroni correction to apply.}
}
\value{
A list of:
\item{n}{A data frame of the number of samples required in each
comparison group to detect an effect corresponding to a given set of
sens/spec values (60/60 to 90/90) and given vector of power values
(default \code{power = seq(0.6, 0.95, 0.05)}).
The corresponding KS-distances of the sens/spec values are also included.}
\item{power}{A data frame containing
the power to detect an effect corresponding to a given set of sens/spec
values (60/60 to 90/90) given vector of sample sizes in each comparison
group (default \code{n_vec = seq(20, 100, 10)}). The corresponding KS-distances
of the sens/spec values are also included.
}
}
\description{
Calculate power and sample size tables for a range of known
powers or sample sizes, and relate them to a sensitivity/specificity.
The calculation relies on a conversion from sens/spec to KS-distance and
an approximation from KS-distance to effect size: \code{qnorm((ks + 1) / 2) * 2}.
From there it is similar to \code{\link[=plotPowerCurves]{plotPowerCurves()}}.
}
\examples{
table <- createPowerTable()   # defaults
lapply(table, head)
}
\seealso{
\code{\link[stats:power.t.test]{stats::power.t.test()}}, \code{\link[=ks2delta]{ks2delta()}}
}
\author{
Stu Field
}
