% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-power-table.R
\name{plot.power_table}
\alias{plot.power_table}
\title{S3 Plot Method for class "power_table"}
\usage{
\method{plot}{power_table}(x, plot_power = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{power_table}, ths result of a call to
\code{\link[=createPowerTable]{createPowerTable()}}, which a named list (length=2) with "n" and "power".}

\item{plot_power}{\code{logical(1)}. Should the \code{power} data frame be plotted?
If \code{FALSE}, the sample sizes are plotted.}

\item{...}{Additional arguments passed to \code{\link[graphics:matplot]{graphics::matplot()}}}
}
\description{
Plot power curves of call to \code{\link[=createPowerTable]{createPowerTable()}},
an object of class "power.table". You can plot either power curves or
required sample size curves. This differs from standard power curves
in that the independent variable (usually effect size) is a
varying sensitivity/specificity value.
}
\examples{
# S3 plot method
plot(table)
plot(table, plot.power = FALSE)
}
\seealso{
\code{\link[graphics:matplot]{graphics::matplot()}}
}
