% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Calculate cube of a number}
\usage{
cube(x)
}
\arguments{
\item{x}{Numeric}
}
\value{
Number
}
\description{
In other words, return the power of three for a given number.
}
\examples{
cube(5)
}
