\name{Distributions}

\alias{Distributions}

\title{Distributions in the PoweR package}

\description{Random variate generation for many standard probability 
  distributions are available in the PoweR package.}

\details{The functions for the random variate generation are named in
  the form \code{law}\eqn{xxxx}.

For the Laplace distribution see \code{\link{law0001.Laplace}}.

For the Normal distribution see \code{\link{law0002.Normal}}.

For the Cauchy distribution see \code{\link{law0003.Cauchy}}.

For the Logistic distribution see \code{\link{law0004.Logistic}}.

For the Gamma distribution see \code{\link{law0005.Gamma}}.

For the Beta distribution see \code{\link{law0006.Beta}}.

For the Uniform distribution see \code{\link{law0007.Uniform}}.

For the Student distribution see \code{\link{law0008.Student}}.

For the Chi-Squared distribution see \code{\link{law0009.Chisquared}}.

For the Log Normal distribution see \code{\link{law0010.LogNormal}}.

For the Weibull distribution see \code{\link{law0011.Weibull}}.

For the Shifted Exponential distribution see \code{\link{law0012.ShiftedExp}}.

For the Power Uniform distribution see \code{\link{law0013.PowerUnif}}.

For the Average Uniform distribution see \code{\link{law0014.AverageUnif}}.

For the UUniform distribution see \code{\link{law0015.UUnif}}.

For the VUniform distribution see \code{\link{law0016.VUnif}}.

For the Johnson SU distribution see \code{\link{law0017.JohnsonSU}}.

For the Tukey distribution see \code{\link{law0018.Tukey}}.

For the Location Contaminated distribution see \code{\link{law0019.LocationCont}}.

For the Johnson SB distribution see \code{\link{law0020.JohnsonSB}}.

For the Skew Normal distribution see \code{\link{law0021.SkewNormal}}.

For the Scale Contaminated distribution see \code{\link{law0022.ScaleCont}}.

For the Generalized Pareto distribution see \code{\link{law0023.GeneralizedPareto}}.

For the Generalized Error distribution see \code{\link{law0024.GeneralizedError}}.

For the Stable distribution see \code{\link{law0025.Stable}}.

For the Gumbel distribution see \code{\link{law0026.Gumbel}}.

For the Frechet distribution see \code{\link{law0027.Frechet}}.

For the Generalized Extreme Value distribution see \code{\link{law0028.GeneralizedExtValue}}.

For the Generalized Arcsine distribution see \code{\link{law0029.GeneralizedArcsine}}.

For the Folded Normal distribution see \code{\link{law0030.FoldedNormal}}.

For the Mixture Normal distribution see \code{\link{law0031.MixtureNormal}}.

For the Truncated Normal distribution see \code{\link{law0032.TruncatedNormal}}.

For the Normal with outliers distribution see \code{\link{law0033.Nout}}.

For the Generalized Exponential Power distribution see \code{\link{law0034.GeneralizedExpPower}}.

For the Exponential distribution see \code{\link{law0035.Exponential}}.

For the Asymmetric Laplace distribution see \code{\link{law0036.AsymmetricLaplace}}.

For the Normal-inverse Gaussian distribution see
\code{\link{law0037.NormalInvGaussian}}.

For the Asymmetric Power Distribution see \code{\link{law0038.AsymmetricPowerDistribution}}.

For the modified Asymmetric Power Distribution see \code{\link{law0039.modifiedAsymmetricPowerDistribution}}.

For the Log-Pareto-tail-normal distribution see \code{\link{law0040.Log-Pareto-tail-normal}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical
Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03
}
\author{P. Lafaye de Micheaux, V. A. Tran
}

\seealso{The CRAN task view on distributions, 
\url{https://CRAN.R-project.org/view=Distributions}, 
mentioning several CRAN packages for additional distributions.
}

\keyword{distributions}
\keyword{datagen}