% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-power.R
\name{fisher_power}
\alias{fisher_power}
\title{Fisher Exact Power}
\usage{
fisher_power(p1, p2, n1, n2, nsim = 200L, alpha = 0.05, ...)
}
\arguments{
\item{p1}{first proportion (group 1)}

\item{p2}{second proportion (group 2)}

\item{n1}{group 1 sample size (column 1)}

\item{n2}{group 2 sample size (column 2)}

\item{nsim}{number of simulations to perform in calculating power.}

\item{alpha}{significance level for p-values.}

\item{...}{Passed to \code{\link[=fisher.test]{fisher.test()}}, e.g. \code{alternative}.}
}
\description{
Fisher Exact Power
}
