% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-power.R
\name{fisher_power_curve}
\alias{fisher_power_curve}
\title{Fisher Power}
\usage{
fisher_power_curve(n, p = 0.85, p_diff = -0.1, ...)
}
\arguments{
\item{n}{\code{integer(n)}. A vector of sample sizes per group.}

\item{p}{\code{double(1)}. Binomial probability of group 1 (the base group).}

\item{p_diff}{\code{double(1)} in \link{-1, 1}. The difference from \code{p}
of group 2. Can be negative.}

\item{...}{Passed to \code{\link[=fisher_sim]{fisher_sim()}}, e.g. \code{nsim}.}
}
\description{
Fisher Power
}
