% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{\%colar\%}
\alias{\%colar\%}
\title{Concatenating strings}
\usage{
vetor1 %colar% vetor2
}
\arguments{
\item{vetor1}{character vector}

\item{vetor2}{character vector}
}
\description{

}
\examples{
"string a" \%colar\% "string b"
}
