\name{help.law}

\alias{help.law}

\title{Help Law}

\description{Open directly the documentation for a specified law using its index.}

\usage{
help.law(law.index)
}

\arguments{
  \item{law.index }{law index as given by function \code{\link{getindex}}.}
}

\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{}, --. 
}
\author{P. Lafaye de Micheaux, V. A. Tran}

\seealso{\code{\link{Distributions}} for other standard distributions.}

\keyword{documentation}