\name{help.stat}

\alias{help.stat}

\title{Help Stat}

\description{Open directly the documentation for a specified goodness-of-fit using its index.}

\usage{
help.stat(stat.index)
}

\arguments{
  \item{stat.index }{statistic index as given by function \code{\link{getindex}}.}
}

\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{}, --. 
}
\author{P. Lafaye de Micheaux, V. A. Tran}

\seealso{See \code{\link{Normality.tests}} for goodness-of-fit tests for
  normality. See \code{\link{Laplace.tests}} for goodness-of-fit tests
  for the Laplace distribution. See \code{\link{Uniformity.tests}} for goodness-of-fit tests for uniformity.
}

\keyword{documentation}