% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{ks2delta}
\alias{ks2delta}
\alias{ks_power_table}
\alias{plot.ks_pwr_table}
\title{Convert KS Distance to Effect Size (Delta)}
\usage{
ks2delta(x)

ks_power_table(
  power = seq(0.6, 0.95, 0.05),
  n_vec = seq(20, 100, 10),
  bonferroni = 1
)

\method{plot}{ks_pwr_table}(x, plot_power = TRUE, ...)
}
\arguments{
\item{x}{\code{numeric(n)} in \verb{[0, 1]}. KS-distance(s).
Or a \code{ks_pwr_table} object, from call to \code{\link[=ks_power_table]{ks_power_table()}},
which is a \code{tibble} with \code{n} and \code{power}.}

\item{power}{\code{numeric(n)} in \verb{(0, 1)}. Which level(s)
of power to evaluate the required samples size.}

\item{n_vec}{\code{integer(n)}. Which sample sizes per group
to evaluate the corresponding power.}

\item{bonferroni}{\code{integer(1)}. Optional Bonferroni correction to apply.}

\item{plot_power}{\code{logical(1)}. Should the \code{power} data frame be plotted?
If \code{FALSE}, the sample sizes are plotted.}

\item{...}{Additional arguments as requpred by the \code{\link[=plot]{plot()}} generic.}
}
\value{
Effect sizes corresponding to \code{x}.

A list of:
\item{n}{A data frame of the number of samples required in each
comparison group to detect an effect corresponding to a given set of
sens/spec values (60/60 to 90/90) and given vector of power values
(default \code{power = seq(0.6, 0.95, 0.05)}).
The corresponding KS-distances of the sens/spec values are also included.}
\item{power}{A data frame containing
the power to detect an effect corresponding to a given set of sens/spec
values (60/60 to 90/90) given vector of sample sizes in each comparison
group (default \code{n_vec = seq(20, 100, 10)}). The corresponding KS-distances
of the sens/spec values are also included.
}
}
\description{
Convert KS-distance to effect size using the standard
inverse error function conversion: \code{qnorm( (x + 1) / 2) * 2}.

Calculate power and sample size tables for a range of known
powers or sample sizes, and relate them to a sensitivity/specificity.
The calculation relies on a conversion from sens/spec to KS-distance and
an approximation from KS-distance to effect size: \code{qnorm((ks + 1) / 2) * 2}.

Plot power curves of call to \code{\link[=ks_power_table]{ks_power_table()}},
an object of class \code{ks_pwr_table}. You can plot
either power curves or required sample size curves.
This differs from standard power curves
in that the independent variable (usually effect size)
is a varying sensitivity/specificity value.
}
\examples{
ks2delta(0.45)
table <- ks_power_table()
table
# S3 plot method
plot(table)

plot(table, plot_power = FALSE)
}
\seealso{
\code{\link[=qnorm]{qnorm()}}

\code{\link[stats:power.t.test]{stats::power.t.test()}}, \code{\link[=ks2delta]{ks2delta()}}
}
\author{
Stu Field
}
