% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks2delta.R
\name{ks2delta}
\alias{ks2delta}
\title{Convert KS Distance to Effect Size (Delta)}
\usage{
ks2delta(x)
}
\arguments{
\item{x}{A KS-distance value or a vector of KS-distances.}
}
\value{
Effect sizes corresponding to \code{x}.
}
\description{
Convert KS-distance to effect size using the standard
inverse error function conversion: \code{qnorm( (x + 1) / 2) * 2}.
}
\examples{
ks2delta(0.45)
}
\seealso{
\code{\link[=qnorm]{qnorm()}}
}
\author{
Stu Field
}
