\name{moments} 
\alias{moments38}
\alias{moments37}
\alias{moments36}
\alias{moments35}
\alias{moments34}
\alias{moments33}
\alias{moments32}
\alias{moments31}
\alias{moments30}
\alias{moments29}
\alias{moments28}
\alias{moments27}
\alias{moments26}
\alias{moments25}
\alias{moments24}
\alias{moments23}
\alias{moments22}
\alias{moments21}
\alias{moments20}
\alias{moments19}
\alias{moments18}
\alias{moments17}
\alias{moments16}
\alias{moments15}
\alias{moments14}
\alias{moments13}
\alias{moments12}
\alias{moments11}
\alias{moments10}
\alias{moments9}
\alias{moments8}
\alias{moments7}
\alias{moments6}
\alias{moments5}
\alias{moments4}
\alias{moments3}
\alias{moments2}
\alias{moments1}

\title{Expectation and variance.}

\description{Evaluate the expectation and variance of a law.}

\details{
  Use the function by typing:
  
  \code{moments}\eqn{j}(\code{x},\code{par1},\code{par2},\eqn{etc.})

  where \eqn{j} is the index of the law and \code{par1}, \code{par2}, \eqn{etc.} are the
  parameters of law \eqn{j}.

}


\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{}, --. 
}
\author{P. Lafaye de Micheaux, V. A. Tran
}
\keyword{distributions}
