\name{Normality.tests}

\alias{Normality.tests}

\title{Goodness-of-fit tests for normality.}

\description{List of goodness-of-fit tests for normality.}

\details{The statistic tests for normality are named in the form \code{stat}\eqn{xxxx}.

For the Lilliefors statistic see \code{\link{stat0001.Lilliefors}}.

For the Anderson-Darling statistic see \code{\link{stat0002.AndersonDarling}}.

For the 1st Zhang-Wu statistic see \code{\link{stat0003.ZhangWu1}}.

For the 2nd Zhang-Wu statistic see \code{\link{stat0004.ZhangWu2}}.

For the Glen-Leemis-Barr statistic see \code{\link{stat0005.GlenLeemisBarr}}.

For the D'Agostino-Pearson statistic see \code{\link{stat0006.DAgostinoPearson}}.

For the Jarque-Bera statistic see \code{\link{stat0007.JarqueBera}}.

For the Doornik-Hansen statistic see \code{\link{stat0008.DoornikHansen}}.

For the Gel-Gastwirth statistic see \code{\link{stat0009.GelGastwirth}}.

For the 1st Hosking statistic see \code{\link{stat0010.Hosking1}}.

For the 2nd Hosking statistic see \code{\link{stat0011.Hosking2}}.

For the 3rd Hosking statistic see \code{\link{stat0012.Hosking3}}.

For the 4th Hosking statistic see \code{\link{stat0013.Hosking4}}.

For the 1st Bontemps-Meddahi statistic see \code{\link{stat0014.BontempsMeddahi1}}.

For the 2nd Bontemps-Meddahi statistic see \code{\link{stat0015.BontempsMeddahi2}}.

For the Brys-Hubert-Struyf statistic see \code{\link{stat0016.BrysHubertStruyf}}.

For the Bonett-Seier statistic see \code{\link{stat0017.BonettSeier}}.

For the Brys-Hubert-Struyf & Bonett-Seier statistic see \code{\link{stat0018.BrysHubertStruyf-BonettSeier}}.

For the 1st Cabana-Cabana statistic see \code{\link{stat0019.CabanaCabana1}}.

For the 2nd Cabana-Cabana statistic see \code{\link{stat0020.CabanaCabana2}}.

For the Shapiro-Wilk statistic see \code{\link{stat0021.ShapiroWilk}}.

For the Shapiro-Francia statistic see \code{\link{stat0022.ShapiroFrancia}}.

For the Shapiro-Wilk statistic modified by Rahman-Govindarajulu see \code{\link{stat0023.ShapiroWilk-RG}}.

For the D'Agostino statistic see \code{\link{stat0024.DAgostino}}.

For the Filliben statistic see \code{\link{stat0025.Filliben}}.

For the Chen-Shapiro statistic see \code{\link{stat0026.ChenShapiro}}.

For the 1st Zhang statistic see \code{\link{stat0027.ZhangQ}}.

For the 2nd Zhang statistic see \code{\link{stat0034.ZhangQstar}}.

For the 3rd Zhang statistic see \code{\link{stat0028.ZhangQQstar}}.

For the Barrio-Cuesta-Matran-Rodriguez statistic see \code{\link{stat0029.BarrioCuestaMatranRodriguez}}.

For the Coin statistic see \code{\link{stat0030.Coin}}.

For the Epps-Pulley statistic see \code{\link{stat0031.EppsPulley}}.

For the Martinez-Iglewicz statistic see \code{\link{stat0032.MartinezIglewicz}}.

For the Gel-Miao-Gastwirth statistic see \code{\link{stat0033.GelMiaoGastwirth}}.

For the Desgagne-LafayeDeMicheaux-Leblanc statistic see \code{\link{stat0035.DesgagneLafayeDeMicheauxLeblanc-Rn}}.

For the new CS Desgagne-LafayeDeMicheaux statistic see \code{\link{stat0036.DesgagneLafayeDeMicheaux-XAPD}}.

For the new CS Desgagne-LafayeDeMicheaux statistic see \code{\link{stat0037.DesgagneLafayeDeMicheaux-ZEPD}}.

For the Spiegelhalter statistic see \code{\link{stat0041.Spiegelhalter}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03 
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Laplace.tests}} for goodness-of-fit tests for
  the Laplace distribution. See \code{\link{Uniformity.tests}} for goodness-of-fit tests for uniformity.
}

\keyword{documentation}
