% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{not_near}
\alias{not_near}
\title{Difference between two numbers.}
\usage{
not_near(x, y, tol = 1e-08)
}
\arguments{
\item{x}{Numeric vectors to compare}

\item{y}{Numeric vectors to compare}

\item{tol}{Tolerance of comparison.}
}
\description{
This is a safe way of comparing if two vectors of floating point numbers.
}
\examples{
not_near(sqrt(2), sqrt(3))
}
\keyword{comparison,}
\keyword{equality}
