% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Common Parameters in \pkg{power}}
\arguments{
\item{nsim}{\code{integer(1)}. Number of simulations to perform
in estimating power, (\eqn{1 - \beta}).}

\item{alpha}{\code{double(1)}. The significance threshold (\eqn{\alpha}),
the probability of a significant result given the null hypothesis,
i.e. a false positive. If desired, Bonferroni correction should
be implemented here.}

\item{n}{\code{integer(1)}. The sample size.}

\item{n_vec}{\code{integer(n)}. A sequence of sample sizes per group
to evaluate the corresponding power.}

\item{delta}{\code{double(1)}. The effect size.}

\item{delta_vec}{\code{numeric(n)}. A sequence of effect sizes
to evaluate the corresponding power.}

\item{power_vec}{\code{numeric(n)} in \verb{(0, 1)}. A sequence of
power levels to evaluate the required samples size.}

\item{main}{\code{character(1)}. Optional title.}
}
\description{
The parameters below are commonly used throughout
the \pkg{power} package.
}
