% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerCurves.R
\name{plotPowerCurves}
\alias{plotPowerCurves}
\title{Plot Power Curves}
\usage{
plotPowerCurves(
  effect = c(0.5, 2, 0.1),
  power.vec = seq(0.5, 0.9, 0.1),
  alpha = 0.05,
  bonferroni = 1129,
  main = NULL
)
}
\arguments{
\item{effect}{Numeric. Effect sizes as a sequence of values.}

\item{power.vec}{Numeric. Vector of power to plot.}

\item{alpha}{Numeric. Alpha for the text, the probability of a false positive.}

\item{bonferroni}{Integer. Bonferroni multiple testing correction
for the number of tests performed.}

\item{main}{Character. An optional title for the main plot.}
}
\value{
A plot of power curves and the data matrix (of sample sizes) used
to create the plot is also returned (invisibly), with the power values
as the columns and effect sizes as the rows.
}
\description{
Uses \code{\link[stats:power.t.test]{stats::power.t.test()}} to generate power
values and plots of multiple power curves.
}
\examples{
pwr <- plotPowerCurves(effect    = seq(0.5, 2, 0.1),
                       power.vec = seq(0.5, 0.9, 0.1))
pwr
}
\seealso{
\code{\link[stats:power.t.test]{stats::power.t.test()}}, \code{\link[graphics:matplot]{graphics::matplot()}}
}
\author{
Stu Field
}
