% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{package}
\name{Power-package}
\alias{Power}
\alias{Power-package}
\title{Power}
\description{
Power and sample-size calculations for regression models. The package
assumes all outcome and predictor variables are centered (mean zero).
}
\details{
\strong{Linear regression:} For the model \eqn{Y \sim X + Z}, the functions
\code{\link{PowerLinReg}} and \code{\link{SampleSizeLinReg}} compute power
for the Wald test of \eqn{H_0: \beta_X = 0} and the sample size required
to achieve a target power, respectively. Both support univariate or
multivariate X and Z via matrix arguments (\code{var_x}, \code{var_z},
\code{cov_xz}).

\strong{Logistic regression:} \code{\link{PowerLogisticReg}} and
\code{\link{SampleSizeLogisticReg}} provide power and sample size for the
Wald test of \eqn{H_0: \beta_X = 0} in the logistic model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zrmacc/Power}
  \item Report bugs at \url{https://github.com/zrmacc/Power/issues}
}

}
\author{
Zachary R. McCaw
}
\keyword{internal}
