% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{Calculate the \strong{power} \code{p} of a number \code{x}}
\usage{
power(x, p)
}
\arguments{
\item{x}{Numeric: base}

\item{p}{Numeric: power}
}
\value{
Numeric
}
\description{
Calculate the \strong{power} \code{p} of a number \code{x}
}
\examples{
power(3, 2)
}
\keyword{internal}
