% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{PowerLinReg}
\alias{PowerLinReg}
\title{Power Calculation}
\usage{
PowerLinReg(beta_x, cov_xz, n, t1e = 0.05, var_resid, var_x, var_z)
}
\arguments{
\item{beta_x}{True regression coefficient for X, vector.}

\item{cov_xz}{Covariance of of X and Z, matrix.}

\item{n}{Sample size.}

\item{t1e}{Type 1 error level.}

\item{var_resid}{Residual variance of Y|(X, Z), scalar.}

\item{var_x}{Variance of X, matrix.}

\item{var_z}{Variance of Z, matrix.}
}
\value{
Numeric power.
}
\description{
Power for testing \eqn{H_{0}:\beta_{X} = 0} in the linear model
\eqn{Y ~ X + Z}. All variables (X, Y, Z) are assumed centered.
}
