% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerWrapper.R
\name{powerWrapper}
\alias{powerWrapper}
\title{Wrapper to Create Power Simulation Data}
\usage{
powerWrapper(
  par.list = list(n = seq(2, 30, 2), delta = seq(0.5, 2.5, 0.1)),
  delta = 1.2,
  n = 20,
  ...
)
}
\arguments{
\item{par.list}{List. A named list of length 2, each containing a
vector of values corresponding to the desired values to evaluate each of
\code{n} and \code{delta}. Default values correspond to a typical range of
"n" and "delta" observed in clinical data.}

\item{delta}{Default effect size while evaluating \code{n}.}

\item{n}{Default sample size per group while evaluating \code{delta}.}

\item{...}{Arguments passed to \code{\link[=simulatePowerData]{simulatePowerData()}}, typically
\code{nboot}, which \emph{must} be passed.}
}
\value{
A list of length 2, each the result of a call to
\code{\link[=simulatePowerData]{simulatePowerData()}}, and containing both the "sample size"
simulation and the "effect size" simulation.
}
\description{
Wrapper to simplify creating data from \code{\link[=simulatePowerData]{simulatePowerData()}}.
}
\examples{
sims <- powerWrapper(nboot = 25, verbose = TRUE)
lapply(sims, head)
}
\seealso{
\code{\link[=simulatePowerData]{simulatePowerData()}}, \code{\link[=calcEmpPower]{calcEmpPower()}}
}
\author{
Stu Field
}
