% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_anova_balanced.R
\name{pwr_anova_balanced}
\alias{pwr_anova_balanced}
\title{Power and sample size to balanced ANOVA.}
\usage{
pwr_anova_balanced(group_means, mean, sigma, pwr = NULL, n = NULL,
sig_level = 0.05)
}
\arguments{
\item{group_means}{vector where each entry is the mean of the treatment}

\item{mean}{population and general mean}

\item{sigma}{populational standard deviation}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{n}{number of observations in each treatment.}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_anova_balanced} returns a list with the following
components:
\describe{
\item{group_means}{means for each treatment}
\item{sigma}{population standard deviation}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with number of observations
\code{n}, \code{N} is the sample size and  power \code{pwr}}
}
}
\description{
\code{pwr_anova_balanced} computes the power and the sample size to ANOVA.
}
\examples{
# Power
pwr_anova_balanced(group_means  = c(15, 20, 25, 30, 35), mean = 15,
sigma = 5,  pwr = NULL, n = 10, sig_level = 0.05)
# Sample size
pwr_anova_balanced(group_means  = c(15, 20, 25, 30, 35), mean = 15,
sigma = 5,  pwr = 0.95, n = NULL, sig_level = 0.05)
}
\keyword{anova,}
\keyword{balanced}
\keyword{hypothesis}
\keyword{level,}
\keyword{power,}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{testing,}
