% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_anova_unbalanced.R
\name{pwr_anova_unbalanced}
\alias{pwr_anova_unbalanced}
\title{Power and sample size to unbalanced ANOVA.}
\usage{
pwr_anova_unbalanced(group_means, size_means, mean, sigma,
sig_level = 0.05)
}
\arguments{
\item{group_means}{vector where each entry is the mean of the treatment}

\item{size_means}{vector where each entry is the sample size for each
treatment}

\item{mean}{population and general mean}

\item{sigma}{populational standard deviation}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_anova_unbalanced} returns a list with the following
components:
\describe{
\item{group_means}{vector where each entry is the mean of the treatment}
\item{size_means}{vector where each entry is the sample size for each
treatment}
\item{sigma}{population standard deviation}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with number of observations
\code{n}, \code{N} is the sample size and  power \code{pwr}}
}
}
\description{
\code{pwr_anova_unbalanced} computes the power and the sample size to
unbalanced ANOVA.
}
\examples{
# Power
pwr_anova_unbalanced(group_means  = c(21, 14, 15, 25),
size_means = c(4, 5, 6, 4), mean = 21, sigma = sqrt(2), sig_level = 0.05)
}
\keyword{anova}
\keyword{hypothesis}
\keyword{level,}
\keyword{power,}
\keyword{significance}
\keyword{testing,}
\keyword{unbalanced}
