% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_chisq_test_association.R
\name{pwr_chisq_test_association}
\alias{pwr_chisq_test_association}
\title{Power and sample size for chi squared test.}
\usage{
pwr_chisq_test_association(es, nrow, ncol, n = NULL, pwr = NULL,
sig_level = 0.05)
}
\arguments{
\item{es}{effect size \code{w =} \eqn{\sqrt{\frac{\chi^2_{obs}}{n}}}}

\item{nrow}{number of categories of the first variable}

\item{ncol}{number of categories of the second variable}

\item{n}{number of observations (sample size)}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_chisq_test_association} returns a list with the following
components:
\describe{
\item{es}{effect size}
\item{nrow}{number of categories of the first variable}
\item{ncol}{number of categories of the second variable}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size and \code{n}
is the number of pairs}
}
}
\description{
\code{pwr_chisq_test_association} computes the power and the sample size
for the chi squared testing to check association between two qualitative
variables.
}
\details{
In this function, we use the effect size \code{es}
\eqn{=\sqrt{\frac{\chi_{obs}^2}{n}}}, where \eqn{n} is the sample size and
\eqn{\chi^2_{obs}} is the statistics test. \code{es} is a coefficient or
measure of association.

The effect size \code{es} \eqn{=w=\sqrt{\frac{w^2}{n}}} is an strength measure
of association. Usually, we use
\describe{
\item{\code{w=0.1}}{to detect a weak association between two variables}
\item{\code{w=0.3}}{to detect a moderate association between two variables}
\item{\code{w=0.5}}{to detect a strong association between two variables}
}
}
\examples{
# Power
pwr_chisq_test_association(es = 0.5, nrow = 5, ncol = 5,
n = 300, pwr = NULL, sig_level = 0.05)
# Sample size
pwr_chisq_test_association(es = 0.5, nrow = 5, ncol = 5,
n = NULL, pwr = 0.99, sig_level = 0.05)
}
\keyword{association}
\keyword{between}
\keyword{chi-squared}
\keyword{hypothesis}
\keyword{level,}
\keyword{power,}
\keyword{qualitative}
\keyword{sample}
\keyword{significance}
\keyword{size,}
\keyword{test,}
\keyword{testing,}
\keyword{two}
