% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_paired_t_test.R
\name{pwr_paired_t_test}
\alias{pwr_paired_t_test}
\title{Power and sample size for paired t-test for two population.}
\usage{
pwr_paired_t_test(
  sigma_d,
  delta,
  delta0 = 0,
  n = NULL,
  pwr = NULL,
  alternative = "two.sided",
  sig_level = 0.05
)
}
\arguments{
\item{sigma_d}{standard deviation of difference of each pair}

\item{delta}{difference of means}

\item{delta0}{difference of means under null hypothesis}

\item{n}{number of pairs}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_paired_t_test} returns a list with the following
components:
\describe{
\item{sigma_d}{standard deviation of difference of each pair}
\item{delta}{difference of means}
\item{delta0}{dfference of means under null hypothesis}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size and \code{n}
is the number of pairs}
}
}
\description{
\code{pwr_paired_t_test} computes the power and the sample size for testing
difference of means of two associates and normal populations .
}
\examples{
# Power
pwr_paired_t_test(sigma_d = 5, delta = 5, delta0 = 0,
n1 = 10, n2 = 10, pwr = NULL, alternative = "two.sided", sig_level = 0.05)
# Sample size
pwr_paired_t_test(sigma_d = 5, delta = 5, delta0 = 0,
n1 = NULL, n2 = NULL, pwr = 0.99, alternative = "two.sided",
sig_level = 0.05)
}
\keyword{dependent}
\keyword{difference}
\keyword{hypothesis}
\keyword{level,}
\keyword{means,}
\keyword{of}
\keyword{paired}
\keyword{populations,}
\keyword{power,}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{t}
\keyword{test,}
\keyword{testing,}
\keyword{two}
