% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_prop_1pop.R
\name{pwr_prop_1pop}
\alias{pwr_prop_1pop}
\title{Power and sample size for testing proportion.}
\usage{
pwr_prop_1pop(p, p0, n = NULL, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
}
\arguments{
\item{p}{populational proportion}

\item{p0}{proportion under null hypothesis}

\item{n}{number of observations (sample size)}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_prop_1pop} returns a list with the following components:
\describe{
\item{p}{populational proportion}
\item{p0}{proportion under null hypothesis}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size \code{n} and
power \code{pwr}}
}
}
\description{
\code{pwr_prop_1pop} computes the power and the sample size for testing
proportion.
}
\details{
Exactly one of the parameters 'n' and 'pwr' must
be passed as NULL, and that parameter is determined from the other.
Notice that the last one has non-NULL default so NULL must be explicitly
passed if you want to compute it.

This function computes internally the effect size, given the population
proportion and proportion under null hypothesis.
These parameters are required.
}
\examples{
# Power
pwr_prop_1pop(sigma = 10, sigma0 = 20, n = 25, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
# Sample size
pwr_prop_1pop(sigma = 10, sigma0 = 20, n = NULL, pwr = 0.95,
alternative = "two.sided", sig_level = 0.05)
}
\keyword{hypothesis}
\keyword{level,}
\keyword{one}
\keyword{population,}
\keyword{power,}
\keyword{proportion}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{test,}
\keyword{testing,}
\keyword{variable,}
