% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_prop_2pop.R
\name{pwr_prop_2pop}
\alias{pwr_prop_2pop}
\title{Power and sample size for t-test for two proportions.}
\usage{
pwr_prop_2pop(p1, p2, delta0 = 0, n1 = NULL, n2 = NULL, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
}
\arguments{
\item{p1}{proportion for the first population}

\item{p2}{proportion fot the second population}

\item{delta0}{difference of proportions}

\item{n1}{number of observations (sample size) for the first population}

\item{n2}{number of observations (sample size) for the second population}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_prop_2pop} returns a list with the following
components:
\describe{
\item{p1}{proportion for the first population}
\item{p2}{proportion for the second population}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size, \code{n1} for the
first population and \code{n2} for the second population}
}
}
\description{
\code{pwr_prop_2pop} computes the power and the sample size for testing
two proportions.
}
\details{
Exactly one of the parameters samples sizes ('n1' and 'n2')
and 'pwr' must be passed as NULL, and that parameter is determined from
the other. Notice that the last one has non-NULL default so NULL must be
explicitly passed if you want to compute it.

#' The parameters 'p1' and 'p2' are required.
The effect size is computed internally.
}
\examples{
# Power
pwr_prop_2pop(p1 = 0.3, p2 = 0.15, n1 = 10, n2 = 10, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
# Sample size
pwr_prop_2pop(p1 = 0.3, p2 = 0.15, n1 = NULL, n2 = NULL, pwr = 0.99,
alternative = "two.sided", sig_level = 0.05)
}
\keyword{difference}
\keyword{hypothesis}
\keyword{level,}
\keyword{of}
\keyword{populations,}
\keyword{power,}
\keyword{proportions,}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{test,}
\keyword{testing,}
\keyword{two}
\keyword{variables,}
\keyword{z}
