% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_sigma_2pop.R
\name{pwr_sigma_2pop}
\alias{pwr_sigma_2pop}
\title{Power and sample size for F test for variance.}
\usage{
pwr_sigma_2pop(sigma1, sigma2, n1 = NULL, n2 = NULL, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
}
\arguments{
\item{sigma1}{populational standard deviation for the first population}

\item{sigma2}{populational standard deviation for the second population}

\item{n1}{number of observations (sample size) for the first population}

\item{n2}{number of observations (sample size) for the second population}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_sigma_2pop} returns a list with the following components:
\describe{
\item{sigma1}{populational standard deviation}
\item{sigma2}{standard deviation under null hypothesis}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size, \code{n1} for the
first population and \code{n2} for the second population}
}
}
\description{
\code{pwr_sigma_2pop} computes the power and the sample size for testing
to compare variances of two normal population.
}
\details{
Exactly one of the parameters samples sizes ('n1' and 'n2')
and 'pwr' must be passed as NULL, and that parameter is determined from
the other. Notice that the last one has non-NULL default so NULL must be
explicitly passed if you want to compute it.

The parameters 'sigma1 and 'sigma2' are required.
The effect size is computed internally.
}
\examples{
# Power
pwr_sigma_2pop(sigma1 = 10, sigma2 = 20, n1 = 25, n2 = 21, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
# Sample size
pwr_sigma_2pop(sigma = 10, sigma0 = 20, n = NULL, pwr = 0.95,
alternative = "two.sided", sig_level = 0.05)
}
\keyword{F}
\keyword{for}
\keyword{hypothesis}
\keyword{level,}
\keyword{population,}
\keyword{power,}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{test}
\keyword{testing,}
\keyword{two}
\keyword{variables,}
\keyword{variance,}
