% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_t_test_2pop_homo.R
\name{pwr_t_test_2pop_homo}
\alias{pwr_t_test_2pop_homo}
\title{Power and sample size for t-test for two population.}
\usage{
pwr_t_test_2pop_homo(sigma, delta, delta0 = 0, n1 = NULL, n2 = NULL,
pwr = NULL, alternative = "two.sided", sig_level = 0.05)
}
\arguments{
\item{sigma}{populational standard deviation}

\item{delta}{difference of means}

\item{delta0}{difference of means under null hypothesis}

\item{n1}{number of observations (sample size) for the first population}

\item{n2}{number of observations (sample size) for the second population}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_t_test_2pop_homo} returns a list with the following
components:
\describe{
\item{sigma}{populational standard deviation}
\item{delta}{difference of means}
\item{delta0}{difference of means under null hypothesis}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size, \code{n1} for the
first population and \code{n2} for the second population}
}
}
\description{
\code{pwr_t_test_2pop_homo} computes the power and the sample size for
testing difference of means of two normal populations with equal and
unknown variance.
}
\details{
Exactly one of the parameters samples sizes ('n1' and 'n2')
and 'pwr' must be passed as NULL, and that parameter is determined from
the other. Notice that the last one has non-NULL default so NULL must be
explicitly passed if you want to compute it.

#' The parameters 'sigma' and 'delta' are required.
The effect size is computed internally.
}
\examples{
# Power
pwr_t_test_2pop_homo(delta = 5, delta0 = 0, sigma = 2, n1 = 10, n2 = 10,
pwr = NULL, alternative = "two.sided", sig_level = 0.05)
# Sample size
pwr_t_test_2pop_homo(delta = 5, delta0 = 0, sigma = 2, n1 = NULL, n2 = NULL,
pwr = 0.95, alternative = "two.sided", sig_level = 0.05)
}
\keyword{and}
\keyword{deviations,}
\keyword{difference}
\keyword{equal}
\keyword{homoscedastic}
\keyword{hypothesis}
\keyword{level,}
\keyword{means,}
\keyword{of}
\keyword{populations,}
\keyword{power,}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{standard}
\keyword{t}
\keyword{test,}
\keyword{testing,}
\keyword{two}
\keyword{unknown}
\keyword{variables,}
