% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_z_fisher_test.R
\name{pwr_z_fisher_test}
\alias{pwr_z_fisher_test}
\title{Power and sample size to test Pearson's correlation.}
\usage{
pwr_z_fisher_test(rho, rho0, n = NULL, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
}
\arguments{
\item{rho}{Pearson's correlation}

\item{rho0}{Pearson's correlation under null hypothesis}

\item{n}{number of observations (sample size)}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_z_fisher_test} returns a list with the following
components:
\describe{
\item{rho}{Pearson's correlation}
\item{rho0}{Pearson's correlation under null hypothesis}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size \code{n} and
power \code{pwr}}
}
}
\description{
\code{pwr_z_fisher_test} computes the power and the sample size for testing
Pearson's correlation using Fisher's Z transform.
}
\details{
It can be proved that
\eqn{\frac{1}{2} \log\left(\frac{1 + r}{1 - r}\right)}
has asympotc normal distribution with mean
#' \eqn{\frac{1}{2} \log\left(\frac{1 + \rho}{1 - \rho}\right)}
and variance \eqn{\sqrt{\frac{1}{n - 3}}}, where \eqn{n} is the sample size,
\eqn{\rho} is the populational Pearson's correlation, and \eqn{r} is the
sample correlation.

It's require to give the population Pearson's correlation and Pearson's
correlation under null hypothesis.
}
\examples{
# Power
pwr_z_fisher_test(rho = 0.8, rho0 = 0.7, n = 100, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
# Sample size
pwr_z_fisher_test(rho = 0.8, rho0 = 0.7, n = NULL, pwr = 0.99,
alternative = "two.sided", sig_level = 0.05)
}
\keyword{Fisher's}
\keyword{Z}
\keyword{correlation,}
\keyword{hypothesis}
\keyword{level,}
\keyword{power,}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{testing,}
\keyword{transform,}
