% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_z_test_1pop.R
\name{pwr_z_test_1pop}
\alias{pwr_z_test_1pop}
\title{Power and sample size for z test for one population}
\usage{
pwr_z_test_1pop(m, m0, sigma, n = NULL, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
}
\arguments{
\item{m}{populational mean}

\item{m0}{mean under null hypothesis}

\item{sigma}{populational standard deviation}

\item{n}{number of observations (sample size)}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_z_test_1pop} returns a list with the following components:
\describe{
\item{m}{populational mean}
\item{m0}{mean under null hypothesis}
\item{sigma}{populational standard deviation}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size \code{n} and
power \code{pwr}}
}
}
\description{
\code{pwr_z_test_1pop} computes the power and the sample size for testing
mean in a normal variable with known variance.
}
\details{
Exactly one of the parameters 'n' and 'pwr' must
be passed as NULL, and that parameter is determined from the other.
Notice that the last one has non-NULL default so NULL must be explicitly
passed if you want to compute it.

This function computes internally the effect size, given the population
mean, mean under null hypothesis and the population standard deviation.
These three parameters are required.
}
\examples{
# Power
pwr_z_test_1pop(m = 20, m0 = 10, sigma = 10, n = 25, pwr = NULL,
alternative = "two.sided", sig_level = 0.05)
# Sample size
pwr_z_test_1pop(m = 20, m0 = 10, sigma = 10, n = NULL, pwr = 0.95,
alternative = "two.sided", sig_level = 0.05)
}
\keyword{hypothesis}
\keyword{level,}
\keyword{one}
\keyword{population,}
\keyword{power,}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{test,}
\keyword{testing,}
\keyword{variable,}
\keyword{z}
