% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_z_test_2pop.R
\name{pwr_z_test_2pop}
\alias{pwr_z_test_2pop}
\title{Power and sample size for Z-test for two population.}
\usage{
pwr_z_test_2pop(sigma1, sigma2, delta, delta0 = 0,  n1 = NULL,
n2 = NULL, pwr = NULL, alternative = "two.sided", sig_level = 0.05)
}
\arguments{
\item{sigma1}{standard deviation for the first population}

\item{sigma2}{standard deviation for the second population}

\item{delta}{populational difference of means}

\item{delta0}{difference of means under null hypothesis}

\item{n1}{number of observations (sample size) for the first population}

\item{n2}{number of observations (sample size) for the second population}

\item{pwr}{power of test \eqn{1 + \beta} (1 minus type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}

\item{sig_level}{significance level (Type I error probability)}
}
\value{
\code{pwr_z_test_2pop} returns a list with the following components:
\describe{
\item{delta}{populational difference of means}
\item{delta0}{difference of means under null hypothesis}
\item{sigma1}{standard deviation for the first population}
\item{sigma2}{standard deviation for the second population}
\item{sig_level}{significance level}
\item{power_sampleSize}{A \code{tibble} with sample size, \code{n1} for the
first population and \code{n2} for the second population}
}
}
\description{
\code{pwr_z_test_2pop} computes the power and the sample size for testing
difference of means of two normal populations with known variances.
}
\details{
Exactly one of the parameters samples sizes ('n1' and 'n2')
and 'pwr' must be passed as NULL, and that parameter is determined from
the other. Notice that the last one has non-NULL default so NULL must be
explicitly passed if you want to compute it.

The parameters 'sigma1, 'sigma2' and 'delta' are required.
The effect size is computed internally.
}
\examples{
# Power
pwr_z_test_2pop(sigma1 = 5, sigma2 = 10, delta = -16, delta0 = 0,
n1 = 10, n2 = 10, pwr = NULL, alternative = "two.sided",
sig_level = 0.05)
# Sample size
pwr_z_test_2pop(sigma1 = 5, sigma2 = 10, delta = -16, delta0 = 0,
n1 = NULL, n2 = NULL, pwr = 0.99, alternative = "two.sided",
sig_level = 0.05)
}
\keyword{deviations,}
\keyword{difference}
\keyword{hypothesis}
\keyword{known}
\keyword{level,}
\keyword{means,}
\keyword{of}
\keyword{populations,}
\keyword{power,}
\keyword{sample}
\keyword{significance}
\keyword{size}
\keyword{standard}
\keyword{test,}
\keyword{testing,}
\keyword{two}
\keyword{variables,}
\keyword{z}
