\name{Resample-class}
\Rdversion{1.1}
\docType{class}
\alias{Resample-class}

\title{Class "Resample"}
\description{A concise (1-5 lines) description of what the class is.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("Resample", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{n.iter}:}{Object of class \code{"integer"}}
  }
}
\section{Methods}{
No methods defined with class "Resample" in the signature.
}
\author{\email{thomas.fabbro@unibas.ch}}
\seealso{\code{"powCalc"}
}
\examples{
showClass("Resample")
}
\keyword{classes}
