\name{SampleSize-class}
\Rdversion{1.1}
\docType{class}
\alias{SampleSize-class}

\title{Class "SampleSize"}
\description{follow
  }
\section{Objects from the Class}{
none
}
\section{Slots}{
  \describe{
    \item{\code{estimate}:}{Object of class \code{"integer"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "SampleSize" in the signature.
}
\author{\email{fabbrot@uhbs.ch}}
\seealso{none
}
\examples{
showClass("SampleSize")
}
\keyword{classes}
