\name{sampleSize-methods}
\docType{methods}
\alias{sampleSize-methods}
\alias{sampleSize,power-method}
\alias{sampleSize}
\title{Methods for Function sampleSize}
\description{
  Methods for function \code{sampleSize}
}
\section{Methods}{
\describe{

\item{}{describe this method here}
}}
\keyword{methods}

