% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-power-t.R
\name{simulate_power_t}
\alias{simulate_power_t}
\alias{plot.power_sim}
\alias{print.power_sim}
\title{Create Power Simulation Data Frame}
\usage{
simulate_power_t(
  sequence,
  n = NULL,
  delta = NULL,
  nsim = 50L,
  reps = 25L,
  verbose = interactive(),
  ...
)

\method{plot}{power_sim}(x, ..., file = NULL)

\method{print}{power_sim}(x, ...)
}
\arguments{
\item{sequence}{\code{numeric(n)}. A sequence of values to vary the
appropriate variable, either \code{n} or \code{delta}.}

\item{n}{\code{integer(1)}. The value for the number of samples
per group to hold constant.}

\item{delta}{\code{double(1)}. The value for the effect size to hold constant.}

\item{nsim}{\code{integer(1)}. Number of simulations to perform
in estimating \eqn{1 - \beta}.}

\item{reps}{\code{integer(1)}. Number of simulations \emph{per box} to generate,
i.e. number of points within each simulation box.}

\item{verbose}{\code{logical(1)}. Should function be run in verbose mode?}

\item{...}{Additional arguments passed either to \code{\link[=t_test_power]{t_test_power()}},
only \verb{alpha =}, or to the S3 plotting method.}

\item{x}{An object of class \code{power_sim}, the result of a call
to \code{\link[=simulate_power_t]{simulate_power_t()}}.}

\item{file}{Character. Optional file name to save file if desired.
Default is \code{NULL}, the default graphics device.}
}
\value{
A \code{power_sim} class object.

A plot of power simulations.
}
\description{
Create a data frame containing the power for either
a given set of sample size or effect size values (\code{variable}),
while holding sample or effect size (whichever is not
defined in \code{variable}) constant.

S3 print method for "power_sim" objects.

S3 print method for \code{power_sim} objects
}
\examples{
# constant effect size
size <- simulate_power_t(seq(10, 50, 2), delta = 0.66, nsim = 25)
size

# constant sample size
delta <- simulate_power_t(seq(0.5, 2.5, 0.2), n = 10, nsim = 25)
delta
plot(size)

plot(delta)
}
\author{
Stu Field
}
