% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePowerData.R
\name{simulatePowerData}
\alias{simulatePowerData}
\alias{plot.power_sim}
\alias{print.power_sim}
\title{Create Power Simulation Data Frame}
\usage{
simulatePowerData(
  sequence,
  nsim = 25,
  n = NULL,
  delta = NULL,
  verbose = interactive(),
  ...
)

\method{plot}{power_sim}(x, ..., file = NULL)

\method{print}{power_sim}(x, ...)
}
\arguments{
\item{sequence}{Sequence of values to vary the appropriate variable, either
\code{n} or \code{delta}.}

\item{nsim}{Numeric. Number of simulations per boxes to generate,
i.e. number of points within each simulation box.}

\item{n}{Numeric. The value for the number of samples per group to hold constant.}

\item{delta}{Numeric. The value for the effect size to hold constant.}

\item{verbose}{Logical. Should function be run in verbose mode?}

\item{...}{Additional arguments passed either to \code{\link[=calcEmpPower]{calcEmpPower()}},
especially \verb{nboot=}, which \emph{must} be passed here, or to
\code{\link[graphics:boxplot]{graphics::boxplot()}} if calling the generic S3 plot method.}

\item{x}{An object of class \code{power_sim}, the result of a call
to \code{\link[=simulatePowerData]{simulatePowerData()}}.}

\item{file}{Character. Optional file name to save file if desired.
Default is \code{NULL}, the default graphics device.}
}
\value{
An object of class \code{power_sim} which is a list containing
the results of the power simulation.

A plot of power simulations.
}
\description{
Create a data frame containing the power for either a given set of sample
size or effect size values (\code{variable}), while holding sample or effect
size (whichever isn't defined in \code{variable}) constant.

S3 print method for "power_sim" objects.

S3 print method for \code{power_sim} objects
}
\examples{
# constant effect size
# must pass nboot via '...'
size <- simulatePowerData(seq(10, 50, 2), delta = 1.2, nboot = 25)
size

# constant sample size
delta <- simulatePowerData(seq(0.5, 2.5, 0.1), n = 20, nboot = 25)
delta
sims <- powerWrapper(nboot = 20)
plot(sims$n)
}
\seealso{
\code{\link[=calcEmpPower]{calcEmpPower()}}
}
\author{
Stu Field
}
