% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-n.R
\name{solve_n}
\alias{solve_n}
\title{Solve for \code{n} at specific power value.}
\usage{
solve_n(pwr_tbl, pwr)
}
\arguments{
\item{pwr_tbl}{A \code{pwr_tbl} object.}

\item{pwr}{\code{double(1)} in \verb{(0, 1)}.
The power to be interpolated.}
}
\description{
Solve for the sample size (\code{n}) for
a specified power value. This function
effectively inverts the power curve and
efficiently returns the sample size
for a given power.
}
\examples{
solve_n(tbl, 0.85)

solve_n(tbl, 0.75)

solve_n(tbl, 0.55)
}
