% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Calculate square of a number}
\usage{
square(x)
}
\arguments{
\item{x}{Numeric}
}
\value{
Number
}
\description{
In other words, return the power of two for a given number.
}
\examples{
square(5)
}
