\name{stat0001.Lilliefors}
\alias{stat0001.Lilliefors}
\alias{stat0001}

\title{The Lilliefors test for normality}

\description{The Lilliefors test for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03 

Lilliefors, H. (1967), \emph{On the Kolmogorov-Smirnov test for normality with mean and variance unknown}, 
_Journal of the American Statistical Association_, *62*, 399-402.
}
\author{P. Lafaye de Micheaux, V. A. Tran}

\seealso{See package \code{nortest}. See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.}

\keyword{statistic, test, normality, Lilliefors}
