\name{stat0004.ZhangWu2}
\alias{stat0004.ZhangWu2}
\alias{stat0004}

\title{The 2nd Zhang-Wu test for normality}

\description{The 2nd Zhang-Wu test \code{Z_A} for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Zhang, J. and Wu, Y. (2005), Likelihood-ratio tests for normality, 
\emph{Computational Statistics and Data Analysis}, \bold{49}(3), 709--721.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Zhang-Wu}
