\name{stat0008.DoornikHansen}
\alias{stat0008.DoornikHansen}
\alias{stat0008}

\title{The Doornik-Hansen test for normality}

\description{The Doornik-Hansen test for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Doornik, J.A. and Hansen, H. (1994), \emph{An Omnibus Test for Univariate and Multivariate Normality}, 
Working Paper, Nuffield College, Oxford University, U.K.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}

\keyword{statistic, test, normality, Doornik-Hansen}
