\name{stat0013.Hosking4}
\alias{stat0013.Hosking4}
\alias{stat0013}

\title{The 4th Hosking test for normality}

\description{The 4th Hosking test \code{T_{Lmom}^{(3)}} for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Hosking, J.R.M. (1990), L-moments: analysis and estimation of distributions using linear combinations of order statistics, 
\emph{Journal of the Royal Statistical Society, Series B}, \bold{52}, 105--124.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}

\keyword{statistic, test, normality, Hosking}
