\name{stat0015.BontempsMeddahi2}
\alias{stat0015.BontempsMeddahi2}
\alias{stat0015}

\title{The 2nd Bontemps-Meddahi test for normality}

\description{The 2nd Bontemps-Meddahi test \code{BM_{3-6}} for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Bontemps, C. and Meddahi, N. (2005), Testing Normality: A GMM Approach, 
\emph{Journal of Econometrics}, \bold{124}, 149--186.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}

\keyword{statistic, test, normality, Bontemps-Meddahi}
