\name{stat0018.BrysHubertStruyf-BonettSeier}
\alias{stat0018.BrysHubertStruyf-BonettSeier}
\alias{stat0018}

\title{The Brys-Hubert-Struyf & Bonett-Seier test for normality}

\description{The combination test for normality of Brys-Hubert-Struyf & Bonett-Seier is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Brys, G., Hubert, M. and Struyf, A. (2008), Goodness-of-fit tests based on a robust measure of skewness, 
\emph{Computational Statistics}, \bold{23}(3), 429--442.

Bonett, D.G. and Seier, E. (2002), A test of normality with high uniform power, 
\emph{Computational Statistics and Data Analysis}, \bold{40}, 435--445.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{stat0016.BrysHubertStruyf}} for the
  Brys-Hubert-Struyf test. See \code{\link{stat0017.BonettSeier}} for
  the Bonett-Seier test. See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Brys-Hubert-Struyf, Bonett-Seier}
