\name{stat0024.DAgostino}
\alias{stat0024.DAgostino}
\alias{stat0024}

\title{The D'Agostino test for normality}

\description{The D'Agostino test for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software},
\bold{}, --.

D'Agostino, R.B. (1971), An omnibus test of normality for moderate and large size samples, 
\emph{Biometrika}, \bold{58}, 341--348.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, D'Agostino}
