\name{stat0027.ZhangQ}
\alias{stat0027.ZhangQ}
\alias{stat0027}

\title{The 1st Zhang test for normality}

\description{The 1st Zhang test \code{Q} for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Zhang, P (1999), Omnibus test of normality using the Q statistic, 
\emph{Journal of Applied Statistics}, \bold{26}(4), 519--528.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Zhang}
