\name{stat0028.ZhangQQstar}
\alias{stat0028.ZhangQQstar}
\alias{stat0028}

\title{The 3rd Zhang test for normality}

\description{The 3rd Zhang test \code{Q-Q*} for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software},
\bold{}, --.

Zhang, P (1999), Omnibus test of normality using the Q statistic, 
\emph{Journal of Applied Statistics}, \bold{26}(4), 519--528.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Zhang}
