\name{stat0033.GelMiaoGastwirth}
\alias{stat0033.GelMiaoGastwirth}
\alias{stat0033}

\title{The Gel-Miao-Gastwirth test for normality}

\description{The Gel-Miao-Gastwirth test for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Gel, Y.R., Miao, W. and Gastwirth, J.L. (2007), 
Robust directed tests of normality against heavy-tailed alternatives, 
\emph{Computational Statistics and Data Analysis}, \bold{51}, 2734--2746.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Gel-Miao-Gastwirth}
