\name{stat0036.DesgagneLafayeDeMicheaux-XAPD}
\alias{stat0036.DesgagneLafayeDeMicheaux-XAPD}
\alias{stat0036}

\title{The \eqn{X_{APD}} test for normality}

\description{The \eqn{X_{APD}} test for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Desgagne, A. and Lafaye de Micheaux, P.  (2017), A Powerful and
Interpretable Alternative to the Jarque-Bera Test of Normality Based on
2nd-Power Skewness and Kurtosis, using the Rao's score test on the APD
family, \emph{Journal of Applied Statistics}, .

}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Desgagne-LafayeDeMicheaux}
