\name{stat0038.Glen}
\alias{stat0038.Glen}
\alias{stat0038}

\title{The Glen-Leemis-Barr test for the Laplace distribution}

\description{The Glen-Leemis-Barr test is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Glen, A., Leemis, L., and Barr, D. (2001) Order Statistics in Goodness
of Fit Testing,
\emph{IEEE Transactions on Reliability}, \bold{50}, Number 2, pp. 209-213.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Laplace.tests}} for other goodness-of-fit tests
  for the Laplace distribution.
}
\keyword{statistic, test, Laplace}
