\name{stat0040.Rayner2}
\alias{stat0040.Rayner2}
\alias{stat0040}

\title{The Rayner-Best statistic for the Laplace distribution}

\description{The Rayner-Best statistic for the Laplace distribution is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03

Rayner, J. C. W. and Best, D. J. (1989), Smooth Tests of Goodness of Fit,
Oxford University Press, New York.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Laplace.tests}} for other goodness-of-fit tests
  for the Laplace distribution.
}
\keyword{statistic, test, Laplace}
