\name{stat0056.RaynerBest2}
\alias{stat0056.RaynerBest2}
\alias{stat0056}

\title{The 2nd Rayner-Best test for the Laplace distribution}

\description{The 2nd Rayner-Best test \code{V_4} for the Laplace distribution is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software},
\bold{}, --.

Rayner, J. C. W. and Best, D. J. (1989), \emph{Smooth Tests of Goodness of Fit}, 
Oxford University Press, New York.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Laplace.tests}} for other goodness-of-fit tests for the Laplace distribution.
}
\keyword{statistic, test, Laplace, Rayner-Best}
