\name{stat0058.Kundu}
\alias{stat0058.Kundu}
\alias{stat0058}

\title{The Kundu test for the Laplace distribution}

\description{The Kundu test for the Laplace distribution is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Kundu, Debasis (2005), Discriminating between Normal and Laplace distributions, 
\emph{Advances in ranking and selection, multiple comparisons, and reliability}, 65-79,
Stat. Ind. Technol., Birkhauser Boston, Boston, MA. 
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Laplace.tests}} for other goodness-of-fit tests for the Laplace distribution.
}
\keyword{statistic, test, Laplace, Kundu}
