\name{stat0071.QuesenberryMiller}
\alias{stat0071.QuesenberryMiller}
\alias{stat0071}

\title{The Quesenberry-Miller test for uniformity}

\description{The Quesenberry-Miller test \code{Q} for uniformity is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software},
\bold{}, --.

Quesenberry, C. P. and Miller, F. L. Jr. (1977), Power studies of some tests for uniformity, 
\emph{Journal of Statistical Computation and Simulation}, \bold{5}, 169--191.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Uniformity.tests}} for other goodness-of-fit tests for uniformity.
}
\keyword{statistic, test, uniformity, Quesenberry-Miller}
