\name{stat0081.Zhang1}
\alias{stat0081.Zhang1}
\alias{stat0081}

\title{The 1st Zhang test for uniformity}

\description{The 1st Zhang test \code{Z_A} for uniformity is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}
\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Zhang, J. (2002), Powerful goodness-of-fit tests based on the likelihood ratio, 
\emph{Journal of the Royal Statistical Society Series B}, \bold{64}, 281--294.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Uniformity.tests}} for other goodness-of-fit tests for uniformity.
}
\keyword{statistic, test, uniformity, Zhang}
