\name{stat0082.Zhang2}
\alias{stat0082.Zhang2}
\alias{stat0082}

\title{The 2nd Zhang test for uniformity}

\description{The 2nd Zhang test \code{Z_C} for uniformity is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software},
\bold{}, --.

Zhang, J. (2002), Powerful goodness-of-fit tests based on the likelihood ratio, 
\emph{Journal of the Royal Statistical Society Series B}, \bold{64}, 281--294.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Uniformity.tests}} for other goodness-of-fit tests for uniformity.
}
\keyword{statistic, test, uniformity, Zhang}
