\name{stat0084.DesgagneMicheauxLeblanc-nor1}
\alias{stat0084.DesgagneMicheauxLeblanc-nor1}
\alias{stat0084}

\title{The 1st Desgagne-Micheaux-Leblanc test for normality}

\description{The 1st Desgagne-Micheaux-Leblanc test \code{DLnor1} for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Lafaye de Micheaux, P. and Tran, V. A. (2014), PoweR: Reproducible
Research Tool to ease Monte-Carlo Power Simulation Studies for
Goodness-of-fit Tests in R. \emph{Journal of Statistical Software},
\bold{}, --.

Desgagne, A., Lafaye de Micheaux, P. and Leblanc, A. (2009), P-kurtosis and goodness-of-fit tests for normality, 
\emph{unpublished document}.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Desgagne-Micheaux-Leblanc}
