\name{stat0085.DesgagneMicheauxLeblanc-Lap3}
\alias{stat0085.DesgagneMicheauxLeblanc-Lap3}
\alias{stat0085}

\title{The Desgagne-Micheaux-Leblanc test for the Laplace distribution}

\description{The Desgagne-Micheaux-Leblanc test \code{DLLap3} for the Laplace distribution is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Desgagne, A., Lafaye de Micheaux, P. and Leblanc, A., unpublished document.
}
\author{P. Lafaye de Micheaux, V. A. Tran}

\seealso{See package \code{lawstat}. See \code{\link{Laplace.tests}} for other goodness-of-fit tests for the Laplace distribution.
}
\keyword{statistic, test, Laplace, Desgagne-Micheaux-Leblanc}
