\name{testsPureR} 
\alias{stat1}
\alias{stat2}
\alias{stat3}
\alias{stat4}
\alias{stat5}
\alias{stat6}
\alias{stat7}
\alias{stat8}
\alias{stat9}
\alias{stat10}
\alias{stat11}
\alias{stat12}
\alias{stat13}
\alias{stat14}
\alias{stat15}
\alias{stat16}
\alias{stat17}
\alias{stat18}
\alias{stat19}
\alias{stat20}
\alias{stat21}
\alias{stat22}
\alias{stat23}
\alias{stat24}
\alias{stat25}
\alias{stat26}
\alias{stat27}
\alias{stat28}
\alias{stat29}
\alias{stat30}
\alias{stat31}
\alias{stat32}
\alias{stat33}
\alias{stat34}
\alias{stat35}
\alias{stat36}
\alias{stat37}
\alias{stat38}
\alias{stat39}
\alias{stat40}
\alias{stat41}
\alias{stat42}
\alias{stat43}
\alias{stat44}
\alias{stat45}
\alias{stat46}
\alias{stat47}
\alias{stat48}
\alias{stat49}
\alias{stat50}
\alias{stat51}
\alias{stat52}
\alias{stat53}
\alias{stat54}
\alias{stat55}
\alias{stat56}
\alias{stat57}
\alias{stat58}
\alias{stat59}
\alias{stat60}
\alias{stat61}
\alias{stat62}
\alias{stat63}
\alias{stat64}
\alias{stat65}
\alias{stat66}
\alias{stat67}
\alias{stat68}
\alias{stat69}
\alias{stat70}
\alias{stat71}
\alias{stat72}
\alias{stat73}
\alias{stat74}
\alias{stat75}
\alias{stat76}
\alias{stat77}
\alias{stat78}
\alias{stat79}
\alias{stat80}
\alias{stat81}
\alias{stat82}
\alias{stat83}
\alias{stat84}
\alias{stat85}
\alias{stat86}
\alias{stat87}
\alias{stat88}
\alias{stat89}
\alias{stat90}

\title{Computation of test statistic values in pure R.}

\description{Alternate way to compute test statistic values (only) in
  pure R instead of C/C++, for clarity reasons.}

\details{
  Use the function by typing:
  
  \code{stat}\eqn{j}(\code{x},\code{par1},\code{par2},\eqn{etc.})

  where \eqn{j} is the index of the test and \code{par1}, \code{par2}, \eqn{etc.} are the
  parameters of test \eqn{j}, if any.

}


\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03 
}
\author{P. Lafaye de Micheaux, V. A. Tran
}
\keyword{stats}
