% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-test-power.R
\name{t_test_power}
\alias{t_test_power}
\title{t-test Empirical Power}
\usage{
t_test_power(n, delta, nsim = 200L, alpha = 0.05, ...)
}
\arguments{
\item{n}{\code{integer(1)}. The sample size.}

\item{delta}{\code{numeric(1)}. The effect size.}

\item{nsim}{\code{integer(1)}. Number of simulations to perform
in estimating \eqn{1 - \beta}.}

\item{alpha}{\code{double(1)}. The significance threshold (\eqn{\alpha}).}

\item{...}{Passed to the underlying test, i.e. \code{\link[=fisher.test]{fisher.test()}},
\code{\link[=t.test]{t.test()}}, etc.,  e.g. \verb{alternative =}.}
}
\value{
The value of the empirically calculated power (\eqn{1 - \beta}).
}
\description{
Calculates the empirical power of a test through simulated data
for a given effect size of sample size (per group).
}
\examples{
t_test_power(10, 0.5)
}
\seealso{
\code{\link[=t.test]{t.test()}}, \code{\link[=qt]{qt()}}, \code{\link[=rnorm]{rnorm()}}
}
\author{
Stu Field
}
