% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{table_distribution}
\alias{table_distribution}
\title{Distribution table}
\usage{
table_distribution(df, variable)
}
\arguments{
\item{df}{a tibble}

\item{variable}{a scalar character}
}
\value{
a tibble that is distribution table
}
\description{
\code{table_distribution} computes the distribution table for a character,
integer or discrete vector
}
\examples{
n <- 1000
df <- tibble(nome = sample(c('A', 'B', 'C'), size = n, replace = T,
                          prob = c(0.6, 0.2, 0.2)),
           numero = rnorm(n))
table_distribution(df, 'nome')
}
