\name{theta-methods}
\docType{methods}
\alias{theta-methods}
\alias{theta,powPar-method}
\title{Methods for Function theta}
\description{
  Methods for function \code{theta}. Have a look at the documentation of
  \code{theta}.
}
\section{Methods}{
\describe{

\item{}{describe this method here}
}}
\keyword{methods}

