\name{Uniformity.tests}

\alias{Uniformity.tests}

\title{Goodness-of-fit tests for uniformity}

\description{List of goodness-of-fit tests for uniformity.}

\details{The statistic tests for uniformity are named in the form \code{stat}\eqn{xxxx}.

For the Kolmogorov statistic see \code{\link{stat0063.Kolmogorov}}.

For the Cramer-von Mises statistic see \code{\link{stat0064.CramervonMises}}.

For the Anderson-Darling statistic see \code{\link{stat0065.AndersonDarling}}.

For the Durbin statistic see \code{\link{stat0066.Durbin}}.

For the Kuiper statistic see \code{\link{stat0067.Kuiper}}.

For the 1st Hegazy-Green statistic see \code{\link{stat0068.HegazyGreen1}}.

For the 2nd Hegazy-Green statistic see \code{\link{stat0069.HegazyGreen2}}.

For the Greenwood statistic see \code{\link{stat0070.Greenwood}}.

For the Quesenberry-Miller statistic see \code{\link{stat0071.QuesenberryMiller}}.

For the Read-Cressie statistic see \code{\link{stat0072.ReadCressie}}.

For the Moran statistic see \code{\link{stat0073.Moran}}.

For the 1st Cressie statistic see \code{\link{stat0074.Cressie1}}.

For the 2nd Cressie statistic see \code{\link{stat0075.Cressie2}}.

For the Vasicek statistic see \code{\link{stat0076.Vasicek}}.

For the Swartz statistic see \code{\link{stat0077.Swartz}}.

For the Morales statistic see \code{\link{stat0078.Morales}}.

For the Pardo statistic see \code{\link{stat0079.Pardo}}.

For the Marhuenda statistic see \code{\link{stat0080.Marhuenda}}.

For the 1st Zhang statistic see \code{\link{stat0081.Zhang1}}.

For the 2nd Zhang statistic see \code{\link{stat0082.Zhang2}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03 
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for goodness-of-fit tests for
  normality. See \code{\link{Laplace.tests}} for goodness-of-fit tests for the Laplace distribution.
}

\keyword{documentation}
