% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_fisher_test.R
\name{z_fisher_test}
\alias{z_fisher_test}
\title{Fisher's Z test for Pearson's correlation.}
\usage{
z_fisher_test(x, y, rho0 = 0, alternative = "two.sided",
conf_level = 0.95)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{rho0}{Pearson's correlation under null hypothesis.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{conf_level}{confidence level of the interval}
}
\value{
\code{z_fisher_test} return a list with the following components:
\describe{
\item{statistic}{statistic of test}
\item{p_value}{p-value}
\item{alternative}{a character string specifying the alternative hypothesis}
\item{lower_ci}{lower bound in the confidence interval}
\item{upper_ci}{upper bound in the confidence interval}
\item{conf_level}{confidence level of the interval}
}
}
\description{
\code{z_fisher_test} performs test on Pearson's correlation using Fisher's Z
transform.
}
\details{
It can be proved that
\eqn{\frac{1}{2}\ln\left(\frac{1 + r}{1- r} \right)}
has normal distribution with mean
\eqn{\frac{1}{2}\ln\left(\frac{1 + \rho}{1- \rho} \right)}
and variance \eqn{\sigma^2 = \frac{1}{n - 3}}.
}
\examples{
# x and y are not correlated
Sigma <- matrix(c(1, 0, 0, 1), ncol = 2)
m <- MASS::mvrnorm(500, mu = c(0, 1), Sigma = Sigma)
z_fisher_test(m[, 1], m[, 2])
# x and y are correlated
Sigma <- matrix(c(1, 0.85, 0.85, 1), ncol = 2)
m <- MASS::mvrnorm(500, mu = c(0, 1), Sigma = Sigma)
z_fisher_test(m[, 1], m[, 2])
}
