% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPowerTable.R
\name{createPowerTable}
\alias{createPowerTable}
\alias{plot.power_table}
\title{Calculate Power Table}
\usage{
createPowerTable(
  power = seq(0.6, 0.95, 0.05),
  n.vec = seq(20, 100, 10),
  bonferroni = 1129
)

\method{plot}{power_table}(x, plot.power = TRUE, file = NULL, ...)
}
\arguments{
\item{power}{Which level(s) of power to evaluate the required samples size.}

\item{n.vec}{Which sample sizes per group to evaluate the corresponding power.}

\item{bonferroni}{Integer. The Bonferroni correction to apply.}

\item{x}{An object of class \code{power_table}, ths result of a call to
\code{\link[=createPowerTable]{createPowerTable()}}, which a named list (length=2) with "n" and "power".}

\item{plot.power}{Logical. Should the \code{power} data frame be plotted?
If \code{FALSE}, the sample sizes are plotted.}

\item{file}{Character. Optional file path to save plot.}

\item{...}{Additional arguments passed to \code{\link[graphics:matplot]{graphics::matplot()}}}
}
\value{
A list of:
\item{n}{A data frame of the number of samples required in each
comparison group to detect an effect corresponding to a given set of
sens/spec values (60/60 to 90/90) and given vector of power values
(default \code{power = seq(0.6, 0.95, 0.05)}).
The corresponding KS-distances of the sens/spec values are also included.}
\item{power}{A data frame containing
the power to detect an effect corresponding to a given set of sens/spec
values (60/60 to 90/90) given vector of sample sizes in each comparison
group (default \code{n.vec=seq(20, 100, 10)}). The corresponding KS-distances
of the sens/spec values are also included.}
}
\description{
Calculate power and sample size tables for a range of known
powers or sample sizes, and relate them to a sensitivity/specificity.
The calculation relies on a conversion from sens/spec to KS-distance and
an approximation from KS-distance to effect size: \code{qnorm((ks + 1) / 2) * 2}.
From there it is similar to \code{\link[=plotPowerCurves]{plotPowerCurves()}}.

Plot power curves of call to \code{\link[=createPowerTable]{createPowerTable()}},
an object of class "power.table". You can plot either power curves or
required sample size curves. This differs from standard power curves
in that the independent variable (usually effect size) is a
varying sensitivity/specificity value.
}
\examples{
table <- createPowerTable()   # defaults
lapply(table, head)
# S3 plot method
plot(table)
plot(table, plot.power = FALSE)

}
\seealso{
\code{\link[stats:power.t.test]{stats::power.t.test()}}, \code{\link[=ks2delta]{ks2delta()}}

\code{\link[graphics:matplot]{graphics::matplot()}}
}
\author{
Stu Field
}
