% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Cube}
\usage{
cube(x)
}
\arguments{
\item{x}{a (possibly vector-valued) numerical input}
}
\value{
a (possibly vector-valued) numerical output, which is the input squared, component-wise.
}
\description{
The thing cubes a number
}
\examples{
cube(3)
cube(c(3,6,9))
cube(0)

}
