% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-power.R
\name{fisher_power}
\alias{fisher_power}
\title{Fisher Exact Empirical Power}
\usage{
fisher_power(p1, p2, n1, n2, nsim = 200L, alpha = 0.05, ...)
}
\arguments{
\item{p1}{The first proportion (group 1).}

\item{p2}{The second proportion (group 2).}

\item{n1}{The group 1 sample size (column 1).}

\item{n2}{The group 2 sample size (column 2).}

\item{nsim}{\code{integer(1)}. Number of simulations to perform
in estimating power, (\eqn{1 - \beta}).}

\item{alpha}{\code{double(1)}. The significance threshold (\eqn{\alpha}),
the probability of a significant result given the null hypothesis,
i.e. a false positive. If desired, Bonferroni correction should
be implemented here.}

\item{...}{Passed to the underlying test, i.e. \code{\link[=fisher.test]{fisher.test()}},
\code{\link[=t.test]{t.test()}}, etc.,  e.g. \verb{alternative =}.}
}
\description{
Calculates the empirical power of a test through simulated data
for a given effect size of sample size (per group).
}
\examples{
fisher_power(0.8, 0.7, 100, 100)

fisher_power(0.6, 0.7, 50, 60)
}
