% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher-power-curve.R
\name{fisher_power_curve}
\alias{fisher_power_curve}
\alias{plot.fisher_power_curve}
\title{Create a Power Curve for Fisher's Exact}
\usage{
fisher_power_curve(n, p = 0.85, p_diff = -0.1, ...)

\method{plot}{fisher_power_curve}(x, ...)
}
\arguments{
\item{n}{\code{integer(n)}. A vector of sample sizes per group.}

\item{p}{\code{double(1)}. Binomial probability of group 1 (the base group).}

\item{p_diff}{\code{double(1)} in \verb{[-1, 1]}. The difference from \code{p}
of group 2. Can be negative.}

\item{...}{Passed to \code{\link[=fisher_power]{fisher_power()}}, e.g. \code{nsim}.}
}
\description{
Generates data for a power curve where power is empirically
estimated over a range of sample size values.
}
\examples{
tbl <- fisher_power_curve(seq(50, 400, 25))
tbl

}
\seealso{
\code{\link[=t_power_curve]{t_power_curve()}}
}
