\name{help.law}

\alias{help.law}

\title{Help Law}

\description{Open directly the documentation for a specified law using its index.}

\usage{
help.law(law.index)
}

\arguments{
  \item{law.index }{law index as given by function \code{\link{getindex}}.}
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical
Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03
}
\author{P. Lafaye de Micheaux, V. A. Tran}

\seealso{\code{\link{Distributions}} for other standard distributions.}

\keyword{documentation}