\name{help.stat}

\alias{help.stat}

\title{Help Stat}

\description{Open directly the documentation for a specified goodness-of-fit using its index.}

\usage{
help.stat(stat.index)
}

\arguments{
  \item{stat.index }{statistic index as given by function \code{\link{getindex}}.}
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical
Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03
}
\author{P. Lafaye de Micheaux, V. A. Tran}

\seealso{See \code{\link{Normality.tests}} for goodness-of-fit tests for
  normality. See \code{\link{Laplace.tests}} for goodness-of-fit tests
  for the Laplace distribution. See \code{\link{Uniformity.tests}} for goodness-of-fit tests for uniformity.
}

\keyword{documentation}