\name{Laplace.tests}

\alias{Laplace.tests}

\title{Goodness-of-fit tests for the Laplace distribution}

\description{List of goodness-of-fit tests for the Laplace distribution.
}

\details{The statistic tests for the Laplace distribution are named in
  the form \code{stat}\eqn{xxxx}.

For the Glen-Leemis-Barr test see \code{\link{stat0038.Glen}}.

For the 1st Rayner-Best statistic test see \code{\link{stat0039.Rayner1}}.

For the 2nd Rayner-Best statistic test see \code{\link{stat0040.Rayner2}}.

For the Anderson-Darling statistic see \code{\link{stat0042.AndersonDarling}}.

For the Cramer-von Mises statistic see \code{\link{stat0043.CramervonMises}}.

For the Watson statistic see \code{\link{stat0044.Watson}}.

For the Kolmogorov-Smirnov statistic see \code{\link{stat0045.KolmogorovSmirnov}}.

For the Kuiper statistic see \code{\link{stat0046.Kuiper}}.

For the 1st Meintanis statistic with moment estimators see \code{\link{stat0047.Meintanis1MO}}.

For the 1st Meintanis statistic with maximum likelihood estimators see \code{\link{stat0048.Meintanis1ML}}.

For the 2nd Meintanis statistic with moment estimators see \code{\link{stat0049.Meintanis2MO}}.

For the 2nd Meintanis statistic with maximum likelihood estimators see \code{\link{stat0050.Meintanis2ML}}.

For the 1st Choi-Kim statistic see \code{\link{stat0051.ChoiKim1}}.

For the 2nd Choi-Kim statistic see \code{\link{stat0052.ChoiKim2}}.

For the 3rd Choi-Kim statistic see \code{\link{stat0053.ChoiKim3}}.

For the Desgagne-Micheaux-Leblanc statistic see \code{\link{stat0054.DesgagneMicheauxLeblanc-Gn}}.

For the 1st Rayner-Best statistic see \code{\link{stat0055.RaynerBest1}}.

For the 2nd Rayner-Best statistic see \code{\link{stat0056.RaynerBest2}}.

For the Langholz-Kronmal statistic see \code{\link{stat0057.LangholzKronmal}}.

For the Kundu statistic see \code{\link{stat0058.Kundu}}.

For the Gulati statistic see \code{\link{stat0059.Gulati}}.

For the Gel statistic see \code{\link{stat0060.Gel}}.

For the 1st Gonzalez-Estrada and Villasenor test see \code{\link{stat0091.Gonzales1}}.

For the 2nd Gonzalez-Estrada and Villasenor test see \code{\link{stat0092.Gonzales2}}.

For the 1st Hogg test see \code{\link{stat0093.Hogg1}}.

For the 2nd Hogg test see \code{\link{stat0094.Hogg2}}.

For the 3rd Hogg test see \code{\link{stat0095.Hogg3}}.

For the 4th Hogg test see \code{\link{stat0096.Hogg4}}.

For the Rizzo and Haman test see \code{\link{stat0097.Rizzo}}.

}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical
Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03
}
\author{P. Lafaye de Micheaux, V. A. Tran}

\seealso{See \code{\link{Normality.tests}} for goodness-of-fit tests for
  normality. See \code{\link{Uniformity.tests}} for goodness-of-fit tests for uniformity.
}
\keyword{documentation}
