% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-power-curves.R
\name{plot_power_curves}
\alias{plot_power_curves}
\title{Plot Power Curves}
\usage{
plot_power_curves(
  delta_vec = c(0.5, 2, 0.1),
  power_vec = seq(0.5, 0.9, 0.1),
  alpha = 0.05,
  main = NULL
)
}
\arguments{
\item{delta_vec}{\code{numeric(n)}. A sequence of effect sizes
to evaluate the corresponding power.}

\item{power_vec}{\code{numeric(n)} in \verb{(0, 1)}. A sequence of
power levels to evaluate the required samples size.}

\item{alpha}{\code{double(1)}. The significance threshold (\eqn{\alpha}),
the probability of a significant result given the null hypothesis,
i.e. a false positive. If desired, Bonferroni correction should
be implemented here.}

\item{main}{\code{character(1)}. Optional title.}
}
\value{
A \code{ggplot}.
}
\description{
Uses \code{\link[stats:power.t.test]{stats::power.t.test()}} to generate power
values and plots of multiple power curves.
}
\examples{
plot_power_curves(delta_vec = seq(0.5, 2, 0.1),
                  power_vec = seq(0.5, 0.9, 0.1))
}
\seealso{
\code{\link[stats:power.t.test]{stats::power.t.test()}}
}
\author{
Stu Field
}
