% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerCurves.R
\name{plot_power_curves}
\alias{plot_power_curves}
\title{Plot Power Curves}
\usage{
plot_power_curves(
  effect = c(0.5, 2, 0.1),
  power_vec = seq(0.5, 0.9, 0.1),
  alpha = 0.05,
  bonferroni = 1000,
  main = NULL
)
}
\arguments{
\item{effect}{\code{numeric(n)}. Sequence of effect sizes.}

\item{power_vec}{\code{numeric(n)}. Sequence of powers (lines) to plot.}

\item{alpha}{\code{double(1)}. The significance, the probability of a false positive.}

\item{bonferroni}{\code{integer(1)}. Bonferroni multiple testing correction.}

\item{main}{\code{character(1)}. Optional title.}
}
\value{
A plot.
}
\description{
Uses \code{\link[stats:power.t.test]{stats::power.t.test()}} to generate power
values and plots of multiple power curves.
}
\examples{
plot_power_curves(effect = seq(0.5, 2, 0.1), power_vec = seq(0.5, 0.9, 0.1))
}
\seealso{
\code{\link[stats:power.t.test]{stats::power.t.test()}}, \code{\link[graphics:matplot]{graphics::matplot()}}
}
\author{
Stu Field
}
