% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-package.R
\docType{package}
\name{power-package}
\alias{power}
\alias{power-package}
\title{power: R Package for Basic Power Analysis (2 group)}
\description{
Informal suite of functions to calculate 2-group comparison power
   calculations via R built in functionality or via simulation.
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
\keyword{package}
