% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-package.R
\docType{package}
\name{power-package}
\alias{power}
\alias{power-package}
\title{power: R Package for Basic Power Analyses}
\description{
Informal suite of functions to calculate simple power calculations via p-value simulation.
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
\keyword{package}
