% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-power-t.R
\name{simulatePowerData}
\alias{simulatePowerData}
\alias{plot.power_sim}
\alias{print.power_sim}
\title{Create Power Simulation Data Frame}
\usage{
simulatePowerData(
  sequence,
  n = NULL,
  delta = NULL,
  reps = 25,
  verbose = interactive(),
  ...
)

\method{plot}{power_sim}(x, ..., file = NULL)

\method{print}{power_sim}(x, ...)
}
\arguments{
\item{sequence}{Sequence of values to vary the
appropriate variable, either \code{n} or \code{delta}.}

\item{n}{\code{integer(1)}. The value for the number of samples
per group to hold constant.}

\item{delta}{Numeric. The value for the effect size to hold constant.}

\item{verbose}{Logical. Should function be run in verbose mode?}

\item{...}{Additional arguments passed either to \code{\link[=calcEmpPower]{calcEmpPower()}},
especially \verb{nsim =}, which \emph{must} be passed here, or to
\code{\link[graphics:boxplot]{graphics::boxplot()}} if calling the generic S3 plot method.}

\item{x}{An object of class \code{power_sim}, the result of a call
to \code{\link[=simulatePowerData]{simulatePowerData()}}.}

\item{file}{Character. Optional file name to save file if desired.
Default is \code{NULL}, the default graphics device.}

\item{nsim}{\code{integer(1)}. Number of simulations per box to generate,
i.e. number of points within each simulation box.}
}
\value{
An object of class \code{power_sim} which is a list containing
the results of the power simulation.

A plot of power simulations.
}
\description{
Create a data frame containing the power for either
a given set of sample size or effect size values (\code{variable}),
while holding sample or effect size (whichever is not
defined in \code{variable}) constant.

S3 print method for "power_sim" objects.

S3 print method for \code{power_sim} objects
}
\examples{
# constant effect size
# must pass nsim via '...'
size <- simulatePowerData(seq(10, 50, 2), delta = 1.2, nsim = 25)
size

# constant sample size
delta <- simulatePowerData(seq(0.5, 2.5, 0.1), n = 20, nsim = 25)
delta
sims <- powerWrapper(nsim = 20)
plot(sims$n)
}
\author{
Stu Field
}
