% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square}
\usage{
square(x)
}
\arguments{
\item{x}{a (possibly vector-valued) numerical input}
}
\value{
a (possibly vector-valued) numerical output, which is the input squared, component-wise.
}
\description{
The thing squares a number
}
\examples{
square(7)
square(c(1,2,3))
square(0)

}
