\name{stat0006.DAgostinoPearson}
\alias{stat0006.DAgostinoPearson}
\alias{stat0006}

\title{The D'Agostino-Pearson test for normality}

\description{The D'Agostino-Pearson for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


D'Agostino, R.B. and Pearson, E.S (1973), 
Tests for Departure from Normality. Empirical Results for the Distributions of b2 and \eqn{\sqrt b1}, 
\emph{Biometrika}, \bold{60}(3), 613--622.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, D'Agostino-Pearson}
