\name{stat0007.JarqueBera}
\alias{stat0007.JarqueBera}
\alias{stat0007}

\title{The Jarque-Bera test for normality}

\description{The Jarque-Bera test for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Jarque, C.M. and Bera, A.K. (1987), A Test for Normality of Observations and Regression Residuals, 
\emph{International Statistical Review}, \bold{50}(2), 163--172.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}

\keyword{statistic, test, normality, Jarque-Bera}
