\name{stat0017.BonettSeier}
\alias{stat0017.BonettSeier}
\alias{stat0017}

\title{The Bonett-Seier test for normality}

\description{The Bonett-Seier test \code{T_w} for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Bonett, D.G. and Seier, E. (2002), A test of normality with high uniform power, 
\emph{Computational Statistics and Data Analysis}, \bold{40}, 435--445.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Bonett-Seier}
