\name{stat0023.ShapiroWilk-RG}
\alias{stat0023.ShapiroWilk-RG}
\alias{stat0023}

\title{The Shapiro-Wilk test for normality modified by Rahman-Govindarajulu}

\description{The Shapiro-Wilk test for normality modified by Rahman-Govindarajulu is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Rahman, M.M. and Govindarajulu, Z. (1997), A modification of the test of Shapiro and Wilk for normality, 
\emph{Journal of Applied Statistics}, \bold{24}(2), 219--236.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{stat0021.ShapiroWilk}} for the Shapiro-Wilk
  test. See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Shapiro-Wilk, Rahman-Govindarajulu}
