\name{stat0029.BarrioCuestaMatranRodriguez}
\alias{stat0029.BarrioCuestaMatranRodriguez}
\alias{stat0029}

\title{The Barrio-Cuesta-Matran-Rodriguez test for normality}

\description{The Barrio-CuestaAlbertos-Matran-Rodriguez test for normality is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Barrio, E. del, Cuesta-Albertos, J., Matran, C. and Rodriguez-Rodriguez, J. (1999), 
Tests of goodness-of-fit based on the L_2-Wasserstein distance,
\emph{The Annals of Statistics}, \bold{27}, 1230--1239. 
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Normality.tests}} for other goodness-of-fit tests for normality.
}
\keyword{statistic, test, normality, Barrio-Cuesta-Matran-Rodriguez}
