\name{stat0068.HegazyGreen1}
\alias{stat0068.HegazyGreen1}
\alias{stat0068}

\title{The 1st Hegazy-Green test for uniformity}

\description{The 1st Hegazy-Green test \code{T_1} for uniformity is used

- to compute its statistic and p-value by calling function \code{\link{statcompute}};

- to compute its quantiles by calling function \code{\link{compquant}} or \code{\link{many.crit}};

- to compute its power by calling function \code{\link{powcomp.fast}} or \code{\link{powcomp.easy}}.
}

\references{
Pierre Lafaye de Micheaux, Viet Anh Tran (2016). PoweR: A
Reproducible Research Tool to Ease Monte Carlo Power Simulation
Studies for Goodness-of-fit Tests in R. \emph{Journal of Statistical Software}, \bold{69(3)}, 1--42. doi:10.18637/jss.v069.i03


Hegazy, Y. A. S. and Green, J. R. (1975), Some new goodness-of-fit tests using order statistics, 
\emph{Applied Statistics}, \bold{24}, 299--308.
}
\author{P. Lafaye de Micheaux, V. A. Tran}
\seealso{See \code{\link{Uniformity.tests}} for other goodness-of-fit tests for uniformity.
}
\keyword{statistic, test, uniformity, Hegazy-Green}
